Lots of changes on the USql side.
Staging table reduced slightly in complexity.  Need to reduce the size of the rows so elected to omit type information in the Data column.  So instead of a SqlMap<string, SqlArray<byte[]>> it is a more straightforward and smaller SqlMap<string, string>.  Types of the Git api responses have been figured out long ago and can be done again via other means if needed.

Fixed an issue where all the Processing scripts were incorrectly deduplicating the rows.  It was picking the first row to arrive instead of the most recent one.  True duplicates this dosn't matter but where some are actually updates to a particular entity over time this is important.

Decided to cut most of the long strings to 127k to fit them in a standard USql String.  Trying to keep the entire string in a byte array made the tables so large that they couldn't be re-serialized back into simple csv files.  This in some cases would prevent sharing to other Data Lake accounts.  The exeption to this is the CommitFile patch info which was left as a byte array of patch string information.